feat (lint): add linting to docker paths
