+++ docker/development/eth-node/eslint.config.js
@@ -1,0 +1,3 @@
+const unlockProtocolConfig = require('@unlock-protocol/eslint-config')
+
+module.exports = [...unlockProtocolConfig]
