+++ docs/src/eslint.config.js
@@ -1,0 +1,2 @@
+const unlockProtocolConfig = require('@unlock-protocol/eslint-config')
+module.exports = [...unlockProtocolConfig]
