--- governance/.eslintrc.js
@@ -1,6 +1,0 @@
-module.exports = {
-  extends: ['@unlock-protocol/eslint-config'],
-  rules: {
-    '@typescript-eslint/no-var-requires': 'off',
-  },
-}
