+++ governance/eslint.config.js
@@ -1,0 +1,9 @@
+const unlockConfig = require('@unlock-protocol/eslint-config')
+module.exports = [
+  ...unlockConfig,
+  {
+    rules: {
+      '@typescript-eslint/no-var-requires': 'off',
+    },
+  },
+]
