+++ governance/package.json
@@ -39,7 +39,7 @@
     "check:oracles": "node ./all_networks run scripts/uniswap/checkOracles.js",
     "check:multisig": "yarn hardhat run scripts/multisig/check-all.js",
     "check:cross-chain": "yarn hardhat run scripts/bridge/check.js",
-    "lint:code": "eslint --resolve-plugins-relative-to ../packages/eslint-config .",
+    "lint:code": "eslint .",
     "lint": "yarn lint:code"
   }
 }
