+++ locksmith/package.json
@@ -15,7 +15,7 @@
     "test:db:migrate": "NODE_ENV=test yarn sequelize db:migrate",
     "test:db:drop": "",
     "test:run": "NODE_ENV=test vitest run --poolOptions.threads.singleThread --coverage ",
-    "lint": "eslint --resolve-plugins-relative-to ../packages/eslint-config --ext .ts,.js src/ __tests__/",
+    "lint": "eslint src/ __tests__/",
     "db:migrate": "yarn sequelize db:migrate",
     "sequelize": "sequelize --config ./config/sequelize.config.js",
     "ci": "yarn lint && yarn test && yarn build",
