+++ package.json
@@ -67,7 +67,7 @@
     "logs:infra": "./scripts/start/logs.sh",
     "stop": "./scripts/stop.sh",
     "nuke": "./scripts/nuke.sh",
-    "lint": "yarn packages --parallel run lint --rule '@typescript-eslint/no-require-imports: warn'",
+    "lint": "yarn packages --parallel run lint",
     "lint:contracts": "yarn workspace @unlock-protocol/smart-contracts lint:contracts --fix",
     "postinstall": "husky install"
   },
