+++ packages/contracts/eslint.config.js
@@ -19,4 +19,7 @@ module.exports = [
       },
     },
   },
+  {
+    ignores: ['src/index.ts'],
+  },
 ]
