--- paywall-app/.eslintrc.js
@@ -1,3 +1,0 @@
-module.exports = {
-  extends: ['@unlock-protocol/eslint-config/next'],
-}
