+++ smart-contracts/eslint.config.js
@@ -1,3 +1,4 @@
+const unlockConfig = require('@unlock-protocol/eslint-config')
 const rulesToIgnore = [
   'no-underscore-dangle',
   'no-param-reassign',
@@ -17,12 +18,14 @@ const rulesToIgnore = [
   'import/extensions',
 ]
 
-module.exports = {
-  extends: ['@unlock-protocol/eslint-config'],
-  ignorePatterns: ['coverage'],
-  rules: {
-    ...rulesToIgnore.reduce((obj, rule) => {
-      return { ...obj, [rule]: 'off' }
-    }, {}),
+module.exports = [
+  ...unlockConfig,
+  {
+    ignorePatterns: ['coverage'],
+    rules: {
+      ...rulesToIgnore.reduce((obj, rule) => {
+        return { ...obj, [rule]: 'off' }
+      }, {}),
+    },
   },
-}
+]
