+++ wedlocks/eslint.config.js
@@ -1,0 +1,2 @@
+const unlockConfig = require('@unlock-protocol/eslint-config')
+module.exports = [...unlockConfig]
